+++ CHANGES.md
@@ -6,6 +6,7 @@ This file includes at least a partial list of the major changes in each version.
 v0.3.3 (in progress)
 --------------------
 
+- added `Lazy.clone` to create a shallow copy of an object or array
 - added `Sequence#sort`, which can accept a 2-arity function for more customized sorting (see [#65](https://github.com/dtao/lazy.js/issues/65))
 - both `Sequence#sort` and `Sequence#sortBy` now accept a `descending` argument
 - added `Sequence#indexBy`